crypto-bigint: refactor `UInt` tests
Splits `UInt`-related tests into individual modules for each operation.
