+++ crypto-bigint/tests/lib.rs
@@ -1,0 +1,3 @@
+//! Integration tests
+
+mod uint;
